# resources/openapi/tabou2-api.json
{
	"openapi": "3.0.3",
	"info": {
		"description": "You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
		"version": "1.0.0",
		"title": "tabou2",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "apiteam@swagger.io"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "/{basePath}",
			"description": "Tabou2 v2 API",
			"variables": {
				"basePath": {
					"default": "/v2"
				}
			}
		}
	],
...
	"paths": {
...
		"/programmes": {
			"post": {
				"summary": "Ajout d'un programme d'aménagement",
				"operationId": "createProgramme",
				"requestBody": {
					"description": "Nouveau programme d'aménagement",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/Programme"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/Programme"
								}
							}
						}
					}
				}
			},
			"put": {
				"summary": "Edition d'un programme d'aménagement",
				"operationId": "updateProgramme",
				"requestBody": {
					"description": "Programme d'aménagement à modifier",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/Programme"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/Programme"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			},
			"get": {
				"summary": "Recherche des programmes",
				"operationId": "searchProgrammes",
				"parameters": [
					{
						"in": "query",
						"name": "nom",
						"description": "Recherche par nom",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "etape",
						"description": "Recherche par étape",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "diffusionRestreinte",
						"description": "Indique si les opérations à diffusion restreinte doivent être affichés si les droits le permettent",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "code",
						"description": "Recherche par code",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "numAds",
						"description": "Recherche par numéro ADS",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "nomOperation",
						"description": "Recherche par nom de l'opération associée",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "natureOperation",
						"description": "Recherche par libellé de la nature associée à l'opération",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "clotureDateDebut",
						"description": "Borne inférieure de recherche par date de cloture",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "clotureDateFin",
						"description": "Borne supérieure de recherche par date de cloture",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "tiers",
						"required": false,
						"description": "Recherche par tiers",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "attributionFonciereAnneeDebut",
						"description": "Borne inférieure de recherche par année d'attribution foncière",
						"required": false,
						"schema": {
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "attributionFonciereAnneeFin",
						"description": "Borne supérieure de recherche par année d'attribution foncière",
						"required": false,
						"schema": {
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "attributionDateDebut",
						"description": "Borne inférieure de recherche par date d'attribution",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "attributionDateFin",
						"description": "Borne supérieure de recherche par date d'attribution",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "commercialisationDateDebut",
						"description": "Borne inférieure de recherche par date de commercialisation",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "commercialisationDateFin",
						"description": "Borne supérieure de recherche par date de commercialisation",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "adsDateDebut",
						"description": "Borne inférieure de recherche par date ADS",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "adsDateFin",
						"description": "Borne supérieure de recherche par date ADS",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "livraisonDateDebut",
						"description": "Borne inférieure de recherche par date de livraison",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "livraisonDateFin",
						"description": "Borne supérieure de recherche par date de livraison",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "annulationDate",
						"description": "Recherche par date d'annulation",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "docDateDebut",
						"description": "Borne inférieure de recherche par date DOC",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "docDateFin",
						"description": "Borne supérieure de recherche par date DOC",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "datDateDebut",
						"description": "Borne inférieure de recherche par date DAT",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "datDateFin",
						"description": "Borne supérieure de recherche par date DAT",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"in": "query",
						"name": "logementsAides",
						"description": "True si recherche sur les logements pour lesquels l'état apporte de l'aide",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"name": "start",
						"in": "query",
						"required": false,
						"description": "Index de début - démarre à 0",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "onlyActive",
						"in": "query",
						"required": false,
						"description": "Indique si seuls les éléments actifs doivent être remontés",
						"schema": {
							"type": "boolean",
							"default": true
						}
					},
					{
						"name": "resultsNumber",
						"in": "query",
						"required": false,
						"description": "Nombre de résultats à remonter",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "orderBy",
						"in": "query",
						"required": false,
						"description": "Nom du champ sur lequel trier les résultats",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "asc",
						"in": "query",
						"required": false,
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					}
				}
			}
		},
		"/programmes/{programmeId}": {
			"get": {
				"summary": "Récupération du détail d'un programme d'aménagement",
				"operationId": "getProgrammeById",
				"parameters": [
					{
						"in": "path",
						"name": "programmeId",
						"description": "Identifiant du programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/Programme"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/programmes/{programmeId}/permis": {
			"get": {
				"summary": "Récupération de la liste des permis de construire d'un programme",
				"operationId": "getPermisByProgrammeId",
				"parameters": [
					{
						"in": "path",
						"name": "programmeId",
						"description": "Identifiant du programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "query",
						"name": "start",
						"description": "Index de début - démarre à 0",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "resultsNumber",
						"description": "Nombre de résultats à remonter",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"description": "Nom du champ sur lequel trier les résultats",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "asc",
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"required": false,
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/programmes/{programmeId}/agapeo": {
			"get": {
				"summary": "Récupération de la liste des données agapeo d'un programme",
				"operationId": "getAgapeoByProgrammeId",
				"parameters": [
					{
						"in": "path",
						"name": "programmeId",
						"description": "Identifiant du programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "query",
						"name": "start",
						"description": "Index de début - démarre à 0",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "resultsNumber",
						"description": "Nombre de résultats à remonter",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"description": "Nom du champ sur lequel trier les résultats",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "asc",
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"required": false,
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/programmes/{programmeId}/tiers": {
			"get": {
				"summary": "Recherche les tiers d'un programme",
				"operationId": "searchTiersByProgrammeId",
				"parameters": [
					{
						"in": "path",
						"name": "programmeId",
						"description": "Identifiant du programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"name": "libelle",
						"in": "query",
						"required": false,
						"description": "Libellé du tiers",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "start",
						"in": "query",
						"required": false,
						"description": "Index de début - démarre à 0",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "resultsNumber",
						"in": "query",
						"required": false,
						"description": "Nombre de résultats à remonter",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "orderBy",
						"in": "query",
						"required": false,
						"description": "Nom du champ sur lequel trier les résultats",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "asc",
						"in": "query",
						"required": false,
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			},
			"post": {
				"summary": "Association d'un tiers à un programme",
				"operationId": "associateTiersToProgramme",
				"parameters": [
					{
						"in": "path",
						"name": "programmeId",
						"description": "Identifiant d'un programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"requestBody": {
					"description": "Association entre le tiers, le type tiers",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/TiersTypeTiers"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/AssociationTiersTypeTiers"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/programmes/{programmeId}/tiers/{associationTiersId}": {
			"put": {
				"summary": "Modification d'un tiers d'un programme ",
				"operationId": "updateTiersByProgrammeId",
				"parameters": [
					{
						"in": "path",
						"name": "programmeId",
						"description": "Identifiant d'un programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "associationTiersId",
						"description": "Identifiant du l'association tiers - type tiers - programe",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"requestBody": {
					"description": "Association entre le tiers, le type tiers",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/TiersTypeTiers"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/AssociationTiersTypeTiers"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			},
			"delete": {
				"summary": "Supprime un tiers d'un programme ",
				"operationId": "deleteTiersFromProgramme",
				"parameters": [
					{
						"in": "path",
						"name": "programmeId",
						"description": "Identifiant d'un programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "associationTiersId",
						"description": "Identifiant du l'association tiers - type tiers - programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
...
		"/programmes/{programmeId}/evenements/{evenementId}": {
			"delete": {
				"summary": "Suppression d'un évènement d'un programme d'aménagement",
				"operationId": "deleteEvenementByProgrammeId",
				"parameters": [
					{
						"in": "path",
						"name": "evenementId",
						"description": "Evènement à supprimer",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "programmeId",
						"description": "Identifiant du programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/Programme"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/programmes/{programmeId}/fiche-suivi": {
			"get": {
				"summary": "Récupération de la fiche de suivi d'un programme",
				"operationId": "downloadProgrammeFicheSuivi",
				"parameters": [
					{
						"in": "path",
						"name": "programmeId",
						"description": "Identifiant du programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"404": {
						"description": "La ressource demandée n'existe pas"
					}
				}
			}
		},
...
	}
}